"Modularized product styles into separate productStyles.ts file"
